ch: use vite
