+++ index.html
@@ -2,18 +2,16 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./css/style.min.css" />
-    <script src="./app.js" defer></script>
-    <title>uwitondanishema Muslim</title>
+    <title>Vite App</title>
   </head>
   <body>
     <header>
       <section>
         <nav>
           <a href="#" class="logo">
-            <img src="./images/logo.svg" alt="Muslim uwitondanishema's logo" />
+            <img src="/logo.svg" alt="Muslim uwitondanishema's logo" />
           </a>
           <ul>
             <li><a href="#work">Work</a></li>
@@ -43,7 +41,7 @@
           <div class="grid-box">
             <img
               class="left"
-              src="./images/kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
+              src="kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
               alt=""
             />
             <div class="right description">
@@ -72,14 +70,14 @@
             </div>
             <img
               class="right"
-              src="./images/kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
+              src="kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
               alt=""
             />
           </div>
           <div class="grid-box">
             <img
               class="left"
-              src="./images/kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
+              src="kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
               alt=""
             />
             <div class="right description">
@@ -108,7 +106,7 @@
             </div>
             <img
               class="right"
-              src="./images/kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
+              src="kelly-sikkema-tk9RQCq5eQo-unsplash.jpg"
               alt=""
             />
           </div>
@@ -522,20 +520,21 @@
           </ul>
         </article>
         <article class="images">
-          <img src="./images/ales-nesetril-Im7lZjxeLhg-unsplash.jpg" alt="" />
+          <img src="ales-nesetril-Im7lZjxeLhg-unsplash.jpg" alt="" />
           <img
             class="bounce"
-            src="./images/ales-nesetril-Im7lZjxeLhg-unsplash.jpg"
+            src="ales-nesetril-Im7lZjxeLhg-unsplash.jpg"
             alt=""
           />
           <img
             class="bounce"
-            src="./images/ales-nesetril-Im7lZjxeLhg-unsplash.jpg"
+            src="ales-nesetril-Im7lZjxeLhg-unsplash.jpg"
             alt=""
           />
         </article>
       </section>
     </main>
     <footer>&copy; If you like creativity you got me.</footer>
+    <script type="module" src="/main.js"></script>
   </body>
 </html>
