+++ package.json
@@ -1,0 +1,17 @@
+{
+  "name": "sharedportfolio",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^4.4.0"
+  },
+  "dependencies": {
+    "sass": "^1.63.6"
+  }
+}
